Modifier pattern form createpost

+++ backend/controllers/post.js
@@ -53,21 +53,34 @@ exports.getOnePost = (req, res, next) => {
 };
 
 exports.modifyPost = (req, res, next) => {
-  Post.findOne({ where: { id: req.params.id } })
-    .then((post) => {
-      if (post.UserId === req.token.userId) {
-        // modif objet
-        const postObject = req.file
-          ? {
-              ...req.body.post,
-              image: `${req.protocol}://${req.get("host")}/images/${
-                req.file.filename
-              }`,
-            }
-          : { ...req.body };
-        if (req.file) {
-          const filename = post.image.split("/images/")[1];
-          fs.unlink(`./images/${filename}`, () => {
+  if (req.body.content.length < 3 || req.body.title.length < 3) {
+    res.status(400).json({ error: "trop court" });
+  } else {
+    Post.findOne({ where: { id: req.params.id } })
+      .then((post) => {
+        if (post.UserId === req.token.userId) {
+          // modif objet
+          const postObject = req.file
+            ? {
+                ...req.body.post,
+                image: `${req.protocol}://${req.get("host")}/images/${
+                  req.file.filename
+                }`,
+              }
+            : { ...req.body };
+          if (req.file) {
+            const filename = post.image.split("/images/")[1];
+            fs.unlink(`./images/${filename}`, () => {
+              Post.update(
+                {
+                  ...postObject,
+                },
+                { where: { id: req.params.id } }
+              )
+                .then(() => res.status(200).json({ message: "Post modifié !" }))
+                .catch((error) => res.status(400).json({ error }));
+            });
+          } else {
             Post.update(
               {
                 ...postObject,
@@ -76,24 +89,15 @@ exports.modifyPost = (req, res, next) => {
             )
               .then(() => res.status(200).json({ message: "Post modifié !" }))
               .catch((error) => res.status(400).json({ error }));
-          });
+          }
         } else {
-          Post.update(
-            {
-              ...postObject,
-            },
-            { where: { id: req.params.id } }
-          )
-            .then(() => res.status(200).json({ message: "Post modifié !" }))
-            .catch((error) => res.status(400).json({ error }));
+          res.status(401).json({
+            error: "Vous ne disposez pas des droits pour modifier ce post !",
+          });
         }
-      } else {
-        res.status(401).json({
-          error: "Vous ne disposez pas des droits pour modifier ce post !",
-        });
-      }
-    })
-    .catch((error) => res.status(404).json({ error: "Post introuvable !" }));
+      })
+      .catch((error) => res.status(404).json({ error: "Post introuvable !" }));
+  }
 };
 
 exports.deletePost = (req, res, next) => {
